+++ flutter_app/.gitignore
@@ -1,16 +1,36 @@
+# Miscellaneous
+*.class
+*.log
+*.pyc
+*.swp
+.DS_Store
 .atom/
-.idea
-.vscode/
-.packages
-.pub/
+.buildlog/
+.history
+.svn/
+
+# IntelliJ related
+*.iml
+*.ipr
+*.iws
+.idea/
+
+# The .vscode folder contains launch configuration and tasks you configure in
+# VS Code which you may wish to be included in version control, so this line
+# is commented out by default.
+#.vscode/
+
+# Flutter/Dart/Pub related
+**/doc/api/
 .dart_tool/
-android/
-build/
-ios/
-packages
 .flutter-plugins
-# Since the current expectation is that clients are using Flutter head,
-# and Dart SDK changes unlock dependencies, pubspec.lock isn't useful.
-# Ignore it since it just creates noise. When there's a system in place
-# for versioning this project relative to Flutter, remove this.
-pubspec.lock
+.packages
+.pub-cache/
+.pub/
+/build/
+
+# Web related
+lib/generated_plugin_registrant.dart
+
+# Exceptions to above rules.
+!/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
