+++ flutter_app/.metadata
@@ -4,5 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: eaa9b47a4ac278a9439468911d2c361a472b114b
+  revision: 0e6cb28dbe3857836a4edb4f3f7fc048006bdb64
   channel: master
+
+project_type: app
