+++ flutter_app/macos/.gitignore
@@ -1,6 +1,6 @@
-xcuserdata
+# Flutter-related
+**/Flutter/ephemeral/
+**/Pods/
 
-# Generated by flutter tooling as needed.
-Flutter/ephemeral/
-# Created by CocoaPods for plugins.
-Pods/
+# Xcode-related
+**/xcuserdata/
