--- flutter_app/macos/Configs/Debug.xcconfig
@@ -1,1 +1,0 @@
-#include "../Flutter/Flutter-Debug.xcconfig"
