--- flutter_app/macos/Configs/Release.xcconfig
@@ -1,1 +1,0 @@
-#include "../Flutter/Flutter-Release.xcconfig"
