+++ flutter_app/macos/Runner/Assets.xcassets/AppIcon.appiconset/Contents.json
