+++ flutter_app/macos/Runner/Assets.xcassets/AppIcon.appiconset/app_icon_1024.png
