+++ flutter_app/macos/Runner/Configs/Debug.xcconfig
@@ -1,0 +1,2 @@
+#include "../../Flutter/Flutter-Debug.xcconfig"
+#include "Warnings.xcconfig"
