+++ flutter_app/macos/Runner/Configs/Release.xcconfig
@@ -1,0 +1,2 @@
+#include "../../Flutter/Flutter-Release.xcconfig"
+#include "Warnings.xcconfig"
